# server/generated/model/v1_find_video_response.go
// Code generated by go-swagger; DO NOT EDIT.

package model

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// V1FindVideoResponse 動画を取得する API のレスポンスです
//
// swagger:model v1FindVideoResponse
type V1FindVideoResponse struct {

	// video
	Video *V1Video `json:"video,omitempty"`
}

// Validate validates this v1 find video response
func (m *V1FindVideoResponse) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateVideo(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *V1FindVideoResponse) validateVideo(formats strfmt.Registry) error {

	if swag.IsZero(m.Video) { // not required
		return nil
	}

	if m.Video != nil {
		if err := m.Video.Validate(formats); err != nil {
			if ve, ok := err.(*errors.Validation); ok {
				return ve.ValidateName("video")
			}
			return err
		}
	}

	return nil
}

// MarshalBinary interface implementation
func (m *V1FindVideoResponse) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *V1FindVideoResponse) UnmarshalBinary(b []byte) error {
	var res V1FindVideoResponse
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}