# server/generated/model/v1_list_video_ids_response.go
// Code generated by go-swagger; DO NOT EDIT.

package model

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// V1ListVideoIdsResponse 動画 YouTubeID 一覧を取得する API のレスポンスです
//
// swagger:model v1ListVideoIdsResponse
type V1ListVideoIdsResponse struct {

	// id
	ID []string `json:"id"`
}

// Validate validates this v1 list video ids response
func (m *V1ListVideoIdsResponse) Validate(formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *V1ListVideoIdsResponse) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *V1ListVideoIdsResponse) UnmarshalBinary(b []byte) error {
	var res V1ListVideoIdsResponse
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}