# server/generated/model/v1_list_videos_response.go
// Code generated by go-swagger; DO NOT EDIT.

package model

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"strconv"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// V1ListVideosResponse 動画一覧を取得する API のレスポンスです
//
// swagger:model v1ListVideosResponse
type V1ListVideosResponse struct {

	// 一覧される動画群です
	Video []*V1Video `json:"video"`
}

// Validate validates this v1 list videos response
func (m *V1ListVideosResponse) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateVideo(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *V1ListVideosResponse) validateVideo(formats strfmt.Registry) error {

	if swag.IsZero(m.Video) { // not required
		return nil
	}

	for i := 0; i < len(m.Video); i++ {
		if swag.IsZero(m.Video[i]) { // not required
			continue
		}

		if m.Video[i] != nil {
			if err := m.Video[i].Validate(formats); err != nil {
				if ve, ok := err.(*errors.Validation); ok {
					return ve.ValidateName("video" + "." + strconv.Itoa(i))
				}
				return err
			}
		}

	}

	return nil
}

// MarshalBinary interface implementation
func (m *V1ListVideosResponse) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *V1ListVideosResponse) UnmarshalBinary(b []byte) error {
	var res V1ListVideosResponse
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}